Ignore swift upload errors if they are caused by a failure to create a container
Swift returns a 403 forbidden whenever a user without the ability
to create a container tries to upload a file (seems stupid, but
whatever). We need to ignore this error and carry on with the
upload iterator to ensure that the upload goes ahead anyway.
If the container exists and we got a 403 for no good reason, the
upload should succeed.